+++ example/basic/metro.config.js
@@ -6,16 +6,16 @@ const root = path.resolve(__dirname, '../..');
 const packages = path.resolve(root, 'packages');
 
 const subpathMap = {
-  '@inox/react-native-edot-sdk/nativeModule': path.resolve(packages, 'react-native/src/nativeModule.ts'),
-  '@inox/react-native-edot-sdk/active-view-context': path.resolve(packages, 'react-native/src/activeViewContext.ts'),
+  '@inoxth/react-native-edot-sdk/nativeModule': path.resolve(packages, 'react-native/src/nativeModule.ts'),
+  '@inoxth/react-native-edot-sdk/active-view-context': path.resolve(packages, 'react-native/src/activeViewContext.ts'),
 };
 
 const config = {
   watchFolders: [root],
   resolver: {
     extraNodeModules: {
-      '@inox/react-native-edot-sdk': path.resolve(packages, 'react-native'),
-      '@inox/react-native-edot-tracer-provider': path.resolve(packages, 'react-native-tracer-provider'),
+      '@inoxth/react-native-edot-sdk': path.resolve(packages, 'react-native'),
+      '@inoxth/react-native-edot-tracer-provider': path.resolve(packages, 'react-native-tracer-provider'),
     },
     nodeModulesPaths: [
       path.resolve(__dirname, 'node_modules'),