+++ example/expo-router/metro.config.js
@@ -6,8 +6,8 @@ const root = path.resolve(__dirname, '../..');
 const packages = path.resolve(root, 'packages');
 
 const subpathMap = {
-  '@inox/react-native-edot-sdk/nativeModule': path.resolve(packages, 'react-native/src/nativeModule.ts'),
-  '@inox/react-native-edot-sdk/active-view-context': path.resolve(packages, 'react-native/src/activeViewContext.ts'),
+  '@inoxth/react-native-edot-sdk/nativeModule': path.resolve(packages, 'react-native/src/nativeModule.ts'),
+  '@inoxth/react-native-edot-sdk/active-view-context': path.resolve(packages, 'react-native/src/activeViewContext.ts'),
 };
 
 const config = getDefaultConfig(__dirname);
@@ -15,9 +15,9 @@ const config = getDefaultConfig(__dirname);
 config.watchFolders = [root];
 
 config.resolver.extraNodeModules = {
-  '@inox/react-native-edot-sdk': path.resolve(root, 'packages/react-native'),
-  '@inox/react-native-edot-navigation': path.resolve(root, 'packages/react-native-navigation'),
-  '@inox/react-native-edot-tracer-provider': path.resolve(root, 'packages/react-native-tracer-provider'),
+  '@inoxth/react-native-edot-sdk': path.resolve(root, 'packages/react-native'),
+  '@inoxth/react-native-edot-navigation': path.resolve(root, 'packages/react-native-navigation'),
+  '@inoxth/react-native-edot-tracer-provider': path.resolve(root, 'packages/react-native-tracer-provider'),
 };
 
 config.resolver.nodeModulesPaths = [