+++ packages/react-native/src/activeViewContext.ts
@@ -1,2 +1,2 @@
-export { ActiveViewContext } from '@inox/react-native-edot-shared';
-export type { ActiveView } from '@inox/react-native-edot-shared';
+export { ActiveViewContext } from '@inoxth/react-native-edot-shared';
+export type { ActiveView } from '@inoxth/react-native-edot-shared';