+++ packages/react-native/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @inox/react-native-edot-sdk
+# @inoxth/react-native-edot-sdk
 
 ## 0.1.0
 
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies [60ae271]
-  - @inox/react-native-edot-shared@0.1.0
+  - @inoxth/react-native-edot-shared@0.1.0